sports: steer the team guys to setup, and pay them in turns not shifts
Eleven of 57 eligible members play tennis, track or hockey. Two of those
schedules arrived; hockey's has not, and may not for weeks.

Colin's instruction: put them on setup, give them extra to make it fair since
setup is the easiest job, but do not dump them on everything.

THE STEER needs no new mechanism. A soft window over 20:00-23:59 and another
over 00:00-12:00, for the whole term, leaves setup as the only untouched type —
a party night needs 20:00-23:59 entire, cleanup needs the morning entire, and
setup is a two-hour block anywhere in 08:00-23:59 across three days. Result:
151 of 172 setup slots, zero night shifts, four cleanups between all eleven.

Blocking the MORNING as well was not the first attempt and is the correction
that mattered. Cleanup is 00:00-12:00 the next day and requires the whole
window, which for somebody with unknown practice times is the WORST shift on
the board, not a soft landing. The first pass gave one sophomore ten cleanups
while trying to give him setup.

THE EXTRA is effort weighting (0020). Setup counts 0.7 of a party night —
two hours in daylight that you choose, against 20:00-23:59 sober at a party.
Fairness is now denominated in effort rather than headcount, so the quota
denominator moved with it. The eleven work ~16 turns against ~8.5 for everyone
else and still land just under quota on effort.

Weighted by the JOB, never the person. "Setup counts 0.7 because it is two
hours in daylight you pick" is a sentence that survives being read out at
chapter. "These eleven get a bigger quota because they play sports" is the same
schedule and an argument.

ONE EXCEPTION, found by measuring rather than by reasoning. Preferences are
ignored for GATED shift types. dj is a party-night window and Mateus De
Vasconcellos plays tennis, so the soft rule sorted him behind the only other
qualified DJ — who then took 42 of 43 nights. That is the same 43-to-0 collapse
the per-type rotation key was written to prevent, arriving through a different
door. A gated pool has no alternative for a preference to yield to, so the turn
count has to win outright. Back to 22/21, with a test.

Also this commit: the CLI can finally write the time and recurrence columns that
have existed since 0012 and been unreachable from every shipped surface, and
`unavailability list` shows them — its absence was a MISREPORT, not a gap, since
"busy Thursdays 6-11pm" rendered identically to "gone September to December".

+++ src/risk/services/policy.py
@@ -106,7 +106,7 @@ def is_senior_in_term(class_year: int | None, *, term_start_year: int, term_is_f
 
 
 def quota_targets(
-    *, rotation_slots: int, senior_count: int, underclass_count: int
+    *, rotation_slots: float, senior_count: int, underclass_count: int
 ) -> tuple[float, float]:
     """Season shift targets as ``(senior_target, underclass_target)``.
